+++ src/java/facade/CompanyInterface.java
@@ -10,7 +10,7 @@ import java.util.List;
 public interface CompanyInterface
 {
     public Company createCompany(Company comp);
-    public Company getCompany (int cvr);
+    public Company getCompany (long cvr);
     public Company updateCompany(Company comp);
     public Company deleteCompany(Company comp);
     public List<Company> getCompanies();
